fix: webpack dev build issue, parser types
