+++ packages/graphiql/resources/webpack.config.js
@@ -14,14 +14,7 @@ const rootPath = (...args) => relPath('../', ...args);
 
 const resultConfig = {
   mode: process.env.NODE_ENV,
-  entry: isHMR
-    ? [
-        'react-hot-loader/patch', // activate HMR for React
-        'webpack-dev-server/client?http://localhost:8080', // bundle the client for webpack-dev-server and connect to the provided endpoint
-        'webpack/hot/only-dev-server', // bundle the client for hot reloading, only- means to only hot reload for successful updates
-        './cdn.ts', // the entry point of our app
-      ]
-    : './cdn.ts',
+  entry: './cdn.ts',
   context: rootPath('src'),
   output: {
     path: rootPath(),
