+++ packages/graphql-language-service-parser/src/index.ts
@@ -14,3 +14,5 @@ export { LexRules, ParseRules, isIgnored } from './Rules';
 export { butNot, list, opt, p, t } from './RuleHelpers';
 
 export { default as onlineParser } from './onlineParser';
+
+export * from './types';
