+++ packages/graphql-language-service-utils/src/index.ts
@@ -12,5 +12,3 @@ export { getASTNodeAtPosition, pointToOffset } from './getASTNodeAtPosition';
 export { Position, Range, locToRange, offsetToPosition } from './Range';
 
 export { validateWithCustomRules } from './validateWithCustomRules';
-
-// export { requireFile, resolveFile } from './file';
